refactor(#5): split deals client into columns, state, and dialog
Extract table constants, filter/format helpers, column defs, column persistence hook, and create dialog so DealsClientContent stays a thin orchestrator.
